Alarm
App that uses various techniques to wake the user up. (Optional) It makes use of the buddy system where you decide an alarm buddy to whom a call will be placed and you can wake each other up. It also uses slumber system to automatically allow user a few minutes to roll around and be lazy in bed(do not use if you want to wake up in a hurry).